\name{char.vector.to.str}
\alias{char.vector.to.str}
\title{converts a vector of chars into a single string or multiple strings, broken by sep}
\usage{
  char.vector.to.str(vec, sep = NULL, collapse = "")
}
\description{
  converts a vector of chars into a single string or
  multiple strings, broken by sep
}

