% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cumsum.ignore}
\alias{cumsum.ignore}
\title{Cummulative sum of number of characters that are ignored}
\usage{
\method{cumsum}{ignore}(ignore)
}
\description{
Cummulative sum of number of characters that are ignored
}

\examples{\dontrun{
  str =c("Now that is a nice matrix. Not?","but short!")
  ignore = rep(FALSE,max(nchar(str)))
  ignore[c(4:5,8:20)] = TRUE
  ignore
  cumsum.ignore(ignore)
  ignore = matrix(ignore,nrow=NROW(str),ncol=length(ignore),byrow=TRUE)
  cumsum.ignore(ignore)
}}
