% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.ignore}
\alias{get.ignore}
\title{Gets a logical ignore vector or list from pos matrices}
\usage{
get.ignore(ignore = NULL, ignore.pos = NULL, only.pos = NULL,
  end = NULL, str = NULL)
}
\description{
Gets a logical ignore vector or list from pos matrices
}

