% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDecimal.R
\name{getDecimal}
\alias{getDecimal}
\title{Decimal Format of Alphanumeric String}
\usage{
getDecimal(string, base = NULL)
}
\arguments{
\item{string}{The string to be converted}

\item{base}{(optional) The base of the input}
}
\value{
The conversion from base \code{base} to base 10 of \code{string}
}
\description{
This function takes an alphanumeric string - a string with only letters or numbers - and returns
the base ten representation of the string when treated as an integer. The base argument attempts
to detect binary, decimal, hex, or full alphanumeric base depending on the input. Alternatively,
the base argument can be specified.
}
\examples{
getDecimal("10101")
getDecimal("1ffa5")
getDecimal("1ffa5",36)
}
