\name{ignore.and.complement.pos}
\alias{ignore.and.complement.pos}
\title{Transforms a boolean vector ignore to a pos matrix and its complement}
\usage{
ignore.and.complement.pos(ignore = NULL, ignore.pos = NULL,
  only.pos = NULL)
}
\value{
A list first element is the pos and complement matrix. The
second element is a vector
}
\description{
Transforms a boolean vector ignore to a pos matrix and its
complement
}

