% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ignore.to.pos}
\alias{ignore.to.pos}
\title{Transforms a boolean vector ignore to a pos matrix}
\usage{
ignore.to.pos(ignore, complement = FALSE)
}
\description{
Transforms a boolean vector ignore to a pos matrix
}

\examples{\dontrun{
  ignore = c(TRUE,FALSE,TRUE,TRUE,FALSE,TRUE,FALSE)
  ignore.to.pos(ignore)
  
  ignore = list(c(TRUE,FALSE,TRUE,TRUE,FALSE,TRUE,FALSE),c(TRUE),c(),c(FALSE,FALSE))
  ignore.to.pos(ignore)
  ignore = c(TRUE,FALSE,TRUE,TRUE,FALSE,TRUE,FALSE)
  ignore.and.complement.pos(ignore)
  
}}
