% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pos.to.ignore}
\alias{pos.to.ignore}
\title{Transforms a pos matrix into a logical ignore vector
Warning length is not}
\usage{
pos.to.ignore(pos, end = 1000000L, complement = FALSE, str = NULL)
}
\description{
Transforms a pos matrix into a logical ignore vector
Warning length is not
}

