% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replace.german.umlaute}
\alias{replace.german.umlaute}
\title{replaces German Umlaute like ö,ü,ä with ascii letters oe, ue, ae}
\usage{
replace.german.umlaute(txt = readLines(file), file = NULL,
  write.file = !is.null(file))
}
\description{
replaces German Umlaute like ö,ü,ä with ascii letters oe, ue, ae
}

