% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sep.lines}
\alias{sep.lines}
\title{transforms a single string with line breaks into a vector with one element for each line}
\usage{
sep.lines(txt, collapse = "\\n")
}
\description{
transforms a single string with line breaks into a vector with one element for each line
}

