% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.ends.with}
\alias{str.ends.with}
\title{Returns als elements of txt that end with pattern}
\usage{
\method{str}{ends.with}(txt, pattern)

\method{str}{ends.with}(txt, pattern)
}
\description{
Returns als elements of txt that end with pattern

Returns als elements of txt that end with pattern
}

