\name{str.extract.all}
\alias{str.extract.all}
\title{Returns a list that contains for each element of str  (or pattern) a vector of all substrings that match the pattern. If for a string no element is matched an empty list is returned}
\usage{
\method{str}{extract.all}(str, pattern, fixed = FALSE, perl = FALSE,
  ignore = NULL)
}
\description{
Returns a list that contains for each element of str (or
pattern) a vector of all substrings that match the pattern.
If for a string no element is matched an empty list is
returned
}

