% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.inpos}
\alias{str.inpos}
\title{Returns a logical vector with TRUE for every character of str that is in pos}
\usage{
\method{str}{inpos}(str, pos)
}
\description{
Returns a logical vector with TRUE for every character of str that is in pos
}

