% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.left.of}
\alias{str.left.of}
\title{Returns the substring left to the first occurence of pattern}
\usage{
\method{str}{left.of}(str, pattern, ..., not.found = str)
}
\description{
Returns the substring left to the first occurence of pattern
}

