\name{str.locate.at.end}
\alias{str.locate.at.end}
\title{Locate a pattern at the end of str}
\usage{
  str.locate.at.end(str, pattern, fixed = TRUE, ...)
}
\description{
  Locate a pattern at the end of str
}

\examples{\dontrun{
  str.locate.at.end(c("0123456012","1230","012","01bsf"),"012")  
}}
