% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.matches.pattern}
\alias{str.matches.pattern}
\title{Check if str completely matches a pattern (not just a substring)}
\usage{
\method{str}{matches.pattern}(str, pattern, fixed = TRUE)
}
\description{
Check if str completely matches a pattern (not just a substring)
}

