% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.number.matches}
\alias{str.number.matches}
\title{Returns the number of matches of pattern in each element of str}
\usage{
\method{str}{number.matches}(str, pattern, ...)
}
\description{
Returns the number of matches of pattern in each element of str
}

