% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.remove.ends}
\alias{str.remove.ends}
\title{remove charcaters on left and right of a string
str.remove.ends(c("ABCDEF","01"),1,3) returns c("BC","")}
\usage{
\method{str}{remove.ends}(str, left = 0, right = 0)
}
\description{
remove charcaters on left and right of a string
str.remove.ends(c("ABCDEF","01"),1,3) returns c("BC","")
}

