% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.replace.list}
\alias{str.replace.list}
\title{Performs sequentially all replacements of pattern and replace on the same strings str}
\usage{
\method{str}{replace.list}(str, pattern, replacement, ...)
}
\description{
A very slow implementation
}

