% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.right.of}
\alias{str.right.of}
\title{Returns the substring right to the first occurence of pattern}
\usage{
\method{str}{right.of}(str, pattern, ..., not.found = str)
}
\description{
Returns the substring right to the first occurence of pattern
}

