\name{str.space}
\alias{str.space}
\title{Returns a string constisting of times spaces, vectorized over times}
\usage{
\method{str}{space}(times, space = " ")

\method{str}{space}(times, space = " ")
}
\description{
Returns a string constisting of times spaces, vectorized
over times

Returns a string constisting of times spaces, vectorized
over times
}

