% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.char.matrix}
\alias{to.char.matrix}
\title{converts into a vector of strings into a matrix of single characters}
\usage{
to.char.matrix(str, drop = FALSE)
}
\description{
converts into a vector of strings into a matrix of single characters
}

