% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.lines}
\alias{merge.lines}
\title{Combines c("A","B") into a single string seperated by line breaks}
\usage{
\method{merge}{lines}(txt, collapse = "\\n")
}
\description{
Combines c("A","B") into a single string seperated by line breaks
}

