% Generated by roxygen2 (4.0.1): do not edit by hand
\name{regexp.fixed}
\alias{regexp.fixed}
\title{strings will be treated as fixed constant in regex}
\usage{
regexp.fixed(str, fixed = TRUE)
}
\arguments{
\item{str}{a string vector}

\item{fixed}{if FALSE just return str}
}
\value{
transformed string vector of same size as str
}
\description{
strings will be treated as fixed constant in regex
}

\examples{\dontrun{
  str = c("A.","*")
  # regexp.fixed transforms strings
  regexp.fixed(str)
  # fixed in stringr flags strings instead
  fixed(str)
}}
