% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.at.pos}
\alias{str.at.pos}
\title{pos is a matrix or a list of matrices specifying positions as returned by str.locate.all}
\usage{
\method{str}{at.pos}(str, pos)
}
\description{
pos is a matrix or a list of matrices specifying positions as returned by str.locate.all
}

