% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.between}
\alias{str.between}
\title{Returns the between the first occurence of start and the first occurence of end}
\usage{
\method{str}{between}(str, start, end, ...)
}
\description{
Returns the between the first occurence of start and the first occurence of end
}

