% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.remove.ignore}
\alias{str.remove.ignore}
\title{ignore is a logical vector or matrix stating which char positions shall be ignored
the function removes the substrings for which ignore=TRUE}
\usage{
\method{str}{remove.ignore}(str, ignore)
}
\description{
ignore is a logical vector or matrix stating which char positions shall be ignored
the function removes the substrings for which ignore=TRUE
}

