\name{str.replace}
\alias{str.replace}
\title{Replaces in str every occurence of pattern by replacement}
\usage{
\method{str}{replace}(str, pattern, replacement, fixed = TRUE, perl = FALSE,
  ignore = NULL, ignore.pos = NULL, only.pos = NULL,
  ignore.pattern = "_IGNORE_", ...)
}
\arguments{
  \item{str}{the string to replaced}

  \item{pattern}{the substring to be replaced}

  \item{replacment}{the new substrings}
}
\value{
a string
}
\description{
Replaces in str every occurence of pattern by replacement
}

