\name{str.space}
\alias{str.space}
\title{Returns a string constisting of times spaces, vectorized over times}
\usage{
  str.space(times, space = " ")

  str.space(times, space = " ")
}
\description{
  Returns a string constisting of times spaces, vectorized
  over times

  Returns a string constisting of times spaces, vectorized
  over times
}

