\name{str.starts.with}
\alias{str.starts.with}
\title{Returns als elements of txt that begin with pattern}
\usage{
  str.starts.with(txt, pattern)
}
\description{
  Returns als elements of txt that begin with pattern
}

\examples{\dontrun{
  str = c("Hi how are you", "hi", "now what", "Hi")
  str.starts.with(str,"Hi")
}}
