% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.starts.with}
\alias{str.starts.with}
\title{Returns als elements of txt that begin with pattern}
\usage{
\method{str}{starts.with}(txt, pattern)
}
\description{
Returns als elements of txt that begin with pattern
}

\examples{\dontrun{
  str = c("Hi how are you", "hi", "now what", "Hi")
  str.starts.with(str,"Hi")
}}
